+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Added
+  * Added inline label example for `terra-input-field`, `terra-form-native-select-field`, `terra-form-combobox-field` and `terra-form-single-select-field`.
   * Added new test example for `terra-form-search-select`
   * Added instruction note for invalid example in `terra-search-select` and `terra-combo-box`.
   * Added accessibility note for selectable button in `terra-button`